+++ Core/Schumix.Irc/Commands/HalfOperator/NickCommand.cs
@@ -38,7 +38,7 @@ namespace Schumix.Irc.Commands
 				return;
 			}
 			
-			SchumixBase.NewNick = true;
+			sIrcBase.Networks[sIRCMessage.ServerName].NewNick = true;
 			string nick = sIRCMessage.Info[4];
 
 			if(!Rfc2812Util.IsValidNick(nick))